Adds structured data extraction via llamaparse
I considered doing this with a model off of deepinfra due to its
extreme cheapness but this is too dang convenient.

I also wanted to use Ruby's instructor lib but the only one that
seems to be maintained is on activesupport 7.x, which won't work
here.

So I just went with llamaparse/cloud's service.

The upshot is that we also now have a nice agent creation class
that is initialized along the way.